remove v6 table
