+++ components/pages/file-repository/index.tsx
@@ -30,7 +30,6 @@ import StatsCard from './StatsCard';
 
 import { FiltersProvider } from './hooks/useFiltersContext';
 import QueryBarContainer from './QueryBar/QueryBarContainer';
-import FileTableV8 from './FileTableV8';
 
 export const PaddedRow = styled(Row)`
   padding-bottom: 8px;
@@ -64,7 +63,6 @@ const RepositoryPage = () => {
               <PaddedRow>
                 <Col xl={12}>
                   <FileTable />
-                  <FileTableV8 />
                 </Col>
               </PaddedRow>
             </ContentBody>
